+++ src/buildings/interfaces/Space.java
@@ -5,6 +5,6 @@ public interface Space
     int getRoomsCount();
     void setRoomsCount(int value);
 
-    int getArea();
-    void setArea(int value);
+    double getArea();
+    void setArea(double value);
 }
